Add files via upload
D&D web scrapper. The idea is to make an application that gets the information from the web and uses it to create an application which allows for the DM to manage monsters spells and information easily. The problem i currently have is that it doesnt seem more useful the scrape the wiki then to use the wiki. I could scrap it to store the data but it would be massive.